+++ packages/get_hooked/lib/src/hooked/_hooked.dart
@@ -1,6 +1,9 @@
+import 'package:meta/meta.dart';
+
 import 'builder.dart';
 import 'renderer.dart';
 
+@internal
 abstract final class Hooked {
   static HookElement? builder;
   static RenderHookElement? renderer;