+++ packages/get_hooked/lib/utils/_all_utils.dart
@@ -2,7 +2,6 @@ export 'animation_view.dart';
 export 'async_controller.dart';
 export 'async_notifier.dart';
 export 'clipped_decorated_box.dart';
-export 'equality_notifier.dart';
 export 'immutable_subtree.dart';
 export 'proxy_notifier.dart';
 export 'record_iterable.dart';